+++ CHANGELOG.rst
@@ -35,9 +35,11 @@ Changed
 -------
 
 - **General**
-    - Upgrade site to Django 1.11.24
+    - Upgrade site to django-sodar-core v0.7.0
+    - Upgrade Python requirements to match django-sodar-core v0.7.0
     - Move graph creation dependencies to ``local_extra.txt`` (#609)
     - Move redis requirement to base.txt (#610)
+    - Include backend Javascript and CSS as implemented in django-sodar-core v0.7.0 (#533)
 - **Samplesheets**
     - Color potentially dangerous links (sodar_core#64)
     - Refactor sheet cell data access and sorting (#597)